Add files via upload
We would be very much happy when someone recognizes and greets us anywhere. What if we have someone to greet us when we enter the campus! We are trying to develop such a recognition system which identifies and greets people. 
Here we take the video captured using the surveillance camera as the input. Initially we use the face dataset that we have collected for training. Face features such as eyes, nose, mouth etc. are extracted using algorithms such as viola-jones. The proportions of the distance between the extracted face features, with the help of a classifier are mapped to identify the corresponding person. The proposed framework consists of four stages. Face Skin Detection (FSD), Facial Features Positioning (FFP), Representative Features Extraction (RFE) and Face Matching (FM) using KNN classifier.  
The proposed system can be used in a wide variety of scenarios such as campus, office etc. after being trained with the corresponding dataset.